renamed package to explicit name the relevance for session
