+++ src/org/tn5250j/sessionsettings/AttributesPanel.java
@@ -23,7 +23,7 @@
  * the Free Software Foundation, Inc., 59 Temple Place, Suite 330,
  * Boston, MA 02111-1307 USA
  */
-package org.tn5250j.settings;
+package org.tn5250j.sessionsettings;
 
 import org.tn5250j.SessionConfig;
 import org.tn5250j.tools.LangTool;
